+++ src/components/content/resource/helpers/notes-entry.tsx
@@ -1,44 +1,52 @@
 import { faChevronRight } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import cx from "classnames";
-import type { Dispatch, SetStateAction } from "react";
+import type { Dispatch, ReactNode, SetStateAction } from "react";
 import { useState } from "react";
 import { DetailEntry, DetailsCard } from "./details-card";
 import { useAnalytics } from "@/components/core/providers/analytics/use-analytics";
 
 export type NotesEntryProps = {
-  details: DetailEntry;
+  details: DetailEntry[];
   hideEmpty?: boolean;
-  id: string;
-  subtitle?: string;
-  title?: string;
+  id?: string;
+  summary?: ReactNode;
   versionId?: string;
   isDetailShownOnOpen?: boolean;
+  documentButton?: ReactNode;
 };
 
 type Props = NotesEntryProps;
 
-export const NotesEntry = ({ details, hideEmpty, title, isDetailShownOnOpen }: Props) => {
-  const [isDetailShown, setIsDetailShown] = useState(isDetailShownOnOpen ?? false);
+export const NotesEntry = ({
+  details,
+  hideEmpty,
+  summary,
+  isDetailShownOnOpen = false,
+  documentButton,
+}: Props) => {
+  const [isDetailShown, setIsDetailShown] = useState(isDetailShownOnOpen);
 
   return (
     <div className="ctw-space-y-1">
       <NoteSummary
-        title={title}
+        summary={summary}
         isDetailShown={isDetailShown}
         setIsDetailShown={setIsDetailShown}
       />
-      {isDetailShown && <DetailsCard details={[details]} hideEmpty={hideEmpty} />}
+      {isDetailShown && (
+        <DetailsCard details={details} hideEmpty={hideEmpty} documentButton={documentButton} />
+      )}
     </div>
   );
 };
 
 export const NoteSummary = ({
-  title,
+  summary,
   isDetailShown,
   setIsDetailShown,
 }: {
-  title?: string;
+  summary?: ReactNode;
   isDetailShown: boolean;
   setIsDetailShown: Dispatch<SetStateAction<boolean>>;
 }) => {
@@ -55,15 +63,8 @@ export const NoteSummary = ({
         return setIsDetailShown(!isDetailShown);
       }}
     >
-      <div className="ctw-flex ctw-items-center ctw-justify-between ctw-rounded-lg ctw-bg-bg-white ctw-p-3 ctw-text-left ctw-outline ctw-outline-1 ctw-outline-bg-dark">
-        <div className="ctw-flex ctw-space-x-3">
-          {title && (
-            <div>
-              <div className="ctw-font-semibold ctw-text-content-black">{title}</div>
-            </div>
-          )}
-          {!title && <div className="ctw-text-content-lighter">Unknown</div>}
-        </div>
+      <div className="ctw-flex ctw-items-center ctw-justify-between ctw-space-x-4 ctw-rounded-lg ctw-bg-bg-white ctw-p-3 ctw-text-left ctw-outline ctw-outline-1 ctw-outline-bg-dark">
+        <div className="ctw-flex ctw-grow ctw-space-x-3">{summary}</div>
         <div className="ctw-flex ctw-items-center ctw-space-x-3">
           <div className="ctw-justify-right ctw-flex">
             <FontAwesomeIcon